+++ vedro/_context.py
@@ -6,4 +6,5 @@ T = TypeVar("T")
 
 
 def context(fn: T) -> T:
+    setattr(fn, "__vedro__context__", True)
     return fn