+++ vedro/core/scenario_discoverer/_create_vscenario.py
@@ -1,4 +1,5 @@
 import inspect
+from pathlib import Path
 from typing import Type
 
 from ..._scenario import Scenario
@@ -8,7 +9,8 @@ from .._virtual_step import VirtualStep
 __all__ = ("create_vscenario",)
 
 
-def create_vscenario(scenario: Type[Scenario]) -> VirtualScenario:
+def create_vscenario(scenario: Type[Scenario], *,
+                     project_dir: Path = Path.cwd()) -> VirtualScenario:
     """
     Create a VirtualScenario from a given Scenario class.
 
@@ -17,6 +19,7 @@ def create_vscenario(scenario: Type[Scenario]) -> VirtualScenario:
     It omits private methods (those starting with '_') and includes only functions.
 
     :param scenario: The Scenario class to convert into a VirtualScenario.
+    :param project_dir: The root directory of the project.
     :return: A VirtualScenario object containing the scenario and its steps.
     """
     steps = []
@@ -27,7 +30,7 @@ def create_vscenario(scenario: Type[Scenario]) -> VirtualScenario:
         if not inspect.isfunction(method):
             continue
         steps.append(VirtualStep(method))
-    return VirtualScenario(scenario, steps)
+    return VirtualScenario(scenario, steps, project_dir=project_dir)
 
 # TODO: Move the create_vscenario method to the ScenarioLoader class in v2.
 #       This will allow the ScenarioLoader to be responsible for creating VirtualScenario objects.